+++ api/bet.js
@@ -28,3 +28,19 @@ export const getRoomBetSlips = async (roomId, paging) => {
     throw error;
   }
 }
+
+export const getSelfBetSlips = async (paging) => {
+  const localAxios = await axiosInstance();
+  try {
+    const response = await localAxios.get(baseURL + "/list/self", {
+      params: {
+        page: paging.page,
+        size: paging.size,
+      },
+    });
+    return response.data;
+  } catch (error) {
+    console.error("Error fetching from backend", error);
+    throw error;
+  }
+}
