--- app/profile/[userId]/invites/page.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const UserInvites = () => {
-  return (
-    <div>UserInvites</div>
-  )
-}
-
-export default UserInvites
