+++ components/authStatus.js
@@ -4,6 +4,7 @@ import { useSession, signIn, signOut } from "next-auth/react";
 import { useEffect } from "react";
 import Loader from "./common/Loader";
 import { useTranslation } from "react-i18next";
+import Link from "next/link";
 
 async function keycloakSessionLogOut() {
   try {
@@ -37,7 +38,9 @@ export default function AuthStatus() {
     } else if (session) {
       return (
         <div className="">
-          <span className="font-bold">{session.username}</span>{" "}
+          <Link href={`/profile/${session.username}`} className="underline">
+            <span className="font-bold">{session.username}</span>{" "}
+          </Link>
           <button
             className="bg-error font-bold text-white py-1 px-2 rounded border border-gray-50"
             onClick={() => {
