+++ i18.js
@@ -19,6 +19,7 @@ i18next
           invite: "Davet Et",
           invites: "Davetler",
           balance: "Bakiye",
+          days: "gün",
           hours: "saat",
           minutes: "dakika",
           before: "önce",
@@ -46,6 +47,7 @@ i18next
           unknown: "Bilinmiyor",
           loginInstead: "Hesabınız var mı? Giriş yapın",
           language: "Dil",
+          betSlips: "Bahis Kuponları",
         },
       },
       en: {
@@ -58,6 +60,7 @@ i18next
           invite: "Invite",
           invites: "Invites",
           balance: "Balance",
+          days: "days",
           hours: "hours",
           minutes: "minutes",
           before: "before",
@@ -85,6 +88,7 @@ i18next
           unknown: "Unknown",
           loginInstead: "Already have an account? Login",
           language: "Language",
+          betSlips: "Bet Slips",
         },
       },
     },
